refactor error into header
